move utils.tsx inside utils folder
