+++ packages/select/src/utils/utils.tsx
@@ -9,8 +9,8 @@ import {
   InternalProps as InternalOptionProps,
   OptionElement,
   ReactEmpty,
-} from './Option';
-import { InternalOptionGroup, OptionGroupElement } from './OptionGroup';
+} from '../Option';
+import { InternalOptionGroup, OptionGroupElement } from '../OptionGroup';
 
 // Any screen smaller than a tablet with no pointer, or a coarse pointer and no hover capability (i.e. touch screen)
 // For more details, see: https://css-tricks.com/touch-devices-not-judged-size/
